Moves `Network::checkpoint_list()` method from `zebra-consensus` to `zebra-chain`.
