+++ zebra-chain/src/parameters/checkpoint.rs
@@ -1,0 +1,4 @@
+//! Types, logic, and constants related to Zebra's list of checkpointed block heights and hashes.
+
+pub mod constants;
+pub mod list;
