+++ zebra-chain/src/parameters/checkpoint/constants.rs
