+++ zebra-node-services/src/lib.rs
@@ -1,6 +1,6 @@
 //! The interfaces of some Zebra node services.
 
-pub mod constants;
+pub use zebra_chain::parameters::checkpoint::constants;
 pub mod mempool;
 
 #[cfg(any(test, feature = "rpc-client"))]
